+++ cmd/cli/kubectl-kyverno/main.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/apply"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/create"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/docs"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/jp"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/oci"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test"
@@ -48,7 +49,14 @@ func enableExperimental() bool {
 }
 
 func registerCommands(cli *cobra.Command) {
-	cli.AddCommand(version.Command(), create.Command(), apply.Command(), test.Command(), jp.Command())
+	cli.AddCommand(
+		apply.Command(),
+		create.Command(),
+		docs.Command(cli),
+		jp.Command(),
+		test.Command(),
+		version.Command(),
+	)
 	if enableExperimental() {
 		cli.AddCommand(oci.Command())
 	}